+++ src/Components/Home.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import NavBar from './NavBar';
+
+function Home() {
+  return (
+    <>
+      <NavBar />
+      <h2 className="title">Welcome to math magicians</h2>
+    </>
+  );
+}
+
+export default Home;
