+++ src/Components/Quotes.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Quotes() {
+  return (
+    <>
+      <h3 className="quote">Here goes thew quotes</h3>
+    </>
+  );
+}
+
+export default Quotes;
