+++ src/Routes.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { BrowserRouter } from 'react-router-dom';
+import App from './Components/App';
+
+const Routes = () => (
+  <BrowserRouter>
+    <App />
+  </BrowserRouter>
+);
+
+export default Routes;
