+++ src/index.js
@@ -2,13 +2,13 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './style/index.scss';
 import './style/media&hover.scss';
-import App from './Components/App';
+import Routes from './Routes';
 import * as serviceWorkerRegistration from './serviceWorkerRegistration';
 import reportWebVitals from './reportWebVitals';
 
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+    <Routes />
   </React.StrictMode>,
   document.getElementById('root'),
 );
